+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Change Log
 
-All notable changes to the "codex-files-by-number" extension will be documented in this file.
+All notable changes to the "codex-manuscript-file-operations" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
 
-- Initial release
+- Initial extension generation